+++ Locales/enUS.lua
@@ -182,6 +182,9 @@ L["Show LOG messages about DEATHS"] = "Show LOG messages about DEATHS";
 L["Show LOG messages about NOT INTERRUPTED casts"] = "Show LOG messages about NOT INTERRUPTED casts";
 L["Options below are global. However they are overridden by LOG-option above. E.g. when \"Show LOG messages\" is disabled, no log messages will be printed independently of settings below."] = "Options below are global. However they are overridden by LOG-option above. E.g. when \"Show LOG messages\" is disabled, no log messages will be printed independently of settings below.";
 L["There is a list of internal trackers. Their info is not used directly on the UI and is useful only for devs. No sense to enable it."] = "There is a list of internal trackers. Their info is not used directly on the UI and is useful only for devs. No sense to enable it.";
+L["Aura"] = "Aura";
+L["MyDungeonsBook %s %s for %s"] = "MyDungeonsBook %s %s for %s"
+L["GUID"] = "GUID";
 -- Settings end
 
 -- BfA start